% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crss_changes_template.R
\name{crss_changes_template}
\alias{crss_changes_template}
\title{Create a template for documenting changes to CRSS}
\usage{
crss_changes_template(filename = NULL, path = NULL)
}
\arguments{
\item{filename}{The name of the file to create. If \code{NULL} will be
crss_changes_template.Rmd. If specified can have a .Rmd extension, and if
it does not, it will be added.}

\item{path}{Where to create the file. If \code{NULL} will be in the current
directory.}
}
\value{
Invisibly returns the full file name if successful. Otherwise returns
\code{FALSE}.
}
\description{
\code{crss_changes_template()} creates a template for documenting the changes to
CRSS since the last release. The template is in RMarkdown and can then be
knit to pdf, and provided with the CRSS package.
}
\examples{
\dontrun{
# create January2020.Rmd in the current directory
crss_changes_template("January2020")
# create crss_changes_template.Rmd in C:/test
crss_changes_template(path = "C:/test")
}

}
