% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_to_excel.R
\name{rdf_to_excel}
\alias{rdf_to_excel}
\title{Convert rdfs to Excel using RdfToExcel}
\usage{
rdf_to_excel(rdfs, path, xlsx = NULL)
}
\arguments{
\item{rdfs}{Vector of rdf files. Should end in ".rdf" and exist in the
specified \code{path}}

\item{path}{Folder containing the rdf files. This is also where the Excel
files will be created.}

\item{xlsx}{The Excel file names. If \code{NULL}, then will use the same file
names as the rdf file names. Otherwise, rename them to names specified
here. If specifying names, this must have the same length as \code{rdfs} and
each name should end in ".xlsx".}
}
\value{
Invisibly returns the absolute path to all created Excel files.
}
\description{
\code{rdf_to_excel()} calls the CADSWES provided RdfToExcel tool for the specified
\code{rdfs}. The RdfToExcelExecutable.exe should be abel to be found in the Path
environment variable, otherwise this function will error.
}
\examples{
\dontrun{
scen_path <- "M:/Shared/CRSS/2020/Scenario/Feb2020_2021,DNF,2007Dems,IG_DCP"
rdfs <- c('KeySlots.rdf','SystemConditions.rdf')
# will save excel files as KeySlots.xlsx and SystemConditions.xlsx
rdf_to_excel(rdfs, scen_path)
# or rename to Feb2020_KeySlots.xlsx and Feb2020_SystemConditions.xlsx
xlsx <- c("Feb2020_KeySlots.xlsx", "Feb2020_SystemConditions.xlsx")
rdf_to_excel(rdfs, scen_path, xlsx)
}

}
