% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_excel_files.R
\name{rename_excel_file}
\alias{rename_excel_file}
\title{Rename Excel Files to Include Scenario Name}
\usage{
rename_excel_file(files, ifolder, insert_name = NULL)
}
\arguments{
\item{files}{Vector of file(s) to rename.}

\item{ifolder}{Top level folder(s) to find the excel files.}

\item{insert_name}{By default (if this is \code{NULL}), the top level folder name
will be inserted into the current file names. If this parameter is
specified, it will be used instead of the top level folder. If it is not
\code{NULL}, it must be the same length as the number of files specified by
\code{ifolder}.}
}
\description{
\code{rename_excel_file()} inserts a scenario name to the beginning of an the
existing file name.
}
