% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_crss_package.R
\name{zip_crss_package}
\alias{zip_crss_package}
\title{Create a zip package for distribution of CRSS.}
\usage{
zip_crss_package(
  ifolder,
  zip_name,
  zip_folders = c("control", "dmi", "model", "results", "ruleset", "sct", "temp")
)
}
\arguments{
\item{ifolder}{Folder to zip}

\item{zip_name}{Name to save the file as. Should end in .zip. Will save this
file in the same directory as ifolder, e.g., if ifolder is
"C:/CRSS/CRSS.Jan2010", then the zip will be saved in "C:/CRSS"}

\item{zip_folders}{The folders to include in the zip.}
}
\description{
\code{zip_crss_package()} creates a zip package that is used to distribute CRSS.
By default, it includes all files in the top level folder, and the following
folders: control, dmi, model, results, ruleset, sct, temp. The dmi folder is
large, and all files are usually not necessary, so will prompt for user input
to determine which folders to include in the dmi folder
}
